Project12/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
django-filter>=23.3
Pillow>=10.0.0
django-redis>=5.4.0
celery>=5.3.0
flower>=2.0.0
python-dateutil>=2.8.2
Faker>=19.0.0
drf-yasg>=1.21.7
gunicorn>=21.0.0
uvicorn>=0.23.0
